Wording for tuple-unpacking via postfix tilde.
For example:

    template<class Vec>
    auto dotproduct(const Vec& a, const Vec& b)
    {
      return ... + (a~ * b~);
    }

For example:

    template<class Idx, class T>
    constexpr decltype(auto) select_elts(Idx indices, T&& values)
    {
      return std::forward_as_tuple(std::get<Idx~>(values)...);
    }

    constexpr auto i = std::make_tuple(3,1,2);
    constexpr auto t = std::make_tuple(5,6,4);
    static_assert(select_elts(i,t) == std::make_tuple(4,5,6));

The wording ends up being almost trivial, at least for value packs!

One problem with the current wording is that it requires the compiler
to instantiate `std::tuple_size_v<decltype(x)>` in order to deal with
`x~`, which means that something as simple as
`[]{ return make_tuple(1,2,3); }~`
won't compile until we get the ability to use lambdas in
unevaluated contexts.

Another problem with the current wording is that it implies that
`std::tuple_size` and `std::get` *must* be instantiated, which
might be user-visible (e.g. if they are then called from another
translation unit that doesn't define them, must we generate code
for them here in order to avoid a user-visible linker error?).
Certainly we don't want to force the compiler to generate
out-of-line code for `std::get<I>`. (Would it usually, though,
just for referring to `std::get<I>`? I would hope not!)

I'd like to include wording for "type packs", but that turns out
to be harder to specify --- I think mainly because I don't have
a good grasp on the grammar for types. The idea would be that,
if `t` is a type-id, then `t~` should be a pack that expands to
`std::tuple_element_t<i,t>`. Type-expressions such as `t~*` and
`int[3]~` should also be well-formed.

For example:

    std::tuple<std::array<int,2>~*...> a;
    static_assert(std::is_same_v<decltype(a), std::tuple<int*,int*>>);