Wording for my proposed LWG 2969 fix.
My impression at the moment is that LWG 2969's fix to
http://eel.is/c++draft/mem.poly.allocator.mem#10.2
was insufficiently similar to the current wording in
http://eel.is/c++draft/allocator.adaptor.members#9.2

The latter talks about constructibility from `inner_allocator_type&`;
the former (after LWG 2969) talks about constructibility from
`polymorphic_allocator` with no ref-qualification. But since we are
perfect-forwarding `*this` through a tuple of references, we definitely
*are* going to try to construct the target object from a `polymorphic_allocator&`
and not from a `polymorphic_allocator`. I believe that the wording in
http://eel.is/c++draft/mem.poly.allocator.mem#10.2
needs to be updated to make the program ill-formed in cases where
that construction is going to fail.

Orthogonally, I believe we need additional `std::move`s in the sentence
following http://eel.is/c++draft/mem.poly.allocator.mem#10.8
for two reasons:

- We don't want to copy-construct the user's arguments that came in
  by-value as part of `x` and/or `y`.

- The value category of the argument to `pair`'s constructor is currently
  unspecified; it could reasonably be either an xvalue or a prvalue.
  Adding the `std::move` makes it clearer that we really want it to be
  an xvalue.