D1144R10, execution policies and ranges
But the Ranges overloads scare me, because you give it two ranges,
and if the output range is shorter then it might not destroy everything
in the input range — but if an exception is thrown then it unquestionably
WILL destroy everything in the input range! That's a bad contract.
So I'm going to remove these again, and document why.